Restructured 'fragments' package
