+++ app/src/main/java/derpibooru/derpy/ui/fragments/tabs/ImageBottomBarFavoritesTabFragment.java
@@ -1,4 +1,4 @@
-package derpibooru.derpy.ui.fragments;
+package derpibooru.derpy.ui.fragments.tabs;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -14,7 +14,7 @@ public class ImageBottomBarFavoritesTabFragment extends ImageBottomBarTabFragmen
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View v = inflater.inflate(R.layout.fragment_image_bottom_bar_favorites_tab, container, false);
+        View v = inflater.inflate(R.layout.tab_image_bottom_bar_favorites, container, false);
         if (getArguments().containsKey("info")) {
             displayInfoInView(v, (DerpibooruImageDetailed) getArguments().getParcelable("info"));
         }
