+++ app/src/main/java/derpibooru/derpy/ui/fragments/tabs/ImageBottomBarTabFragment.java
@@ -1,4 +1,4 @@
-package derpibooru.derpy.ui.fragments;
+package derpibooru.derpy.ui.fragments.tabs;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
