+++ app/src/main/java/derpibooru/derpy/ui/fragments/tabs/SearchOptionsTabFragment.java
@@ -1,4 +1,4 @@
-package derpibooru.derpy.ui.fragments;
+package derpibooru.derpy.ui.fragments.tabs;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -20,7 +20,7 @@ public class SearchOptionsTabFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View v = inflater.inflate(R.layout.fragment_search_options_tab, container, false);
+        View v = inflater.inflate(R.layout.tab_search_options, container, false);
         initSpinnerListeners(v);
         setTextListeners(v);
         return v;
