+++ app/src/main/res/layout/fragment_image_list.xml
