+++ app/src/main/res/layout/tab_image_bottom_bar_info.xml
